app/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3